fulltransnet/requirements.txt
h5py>=3.9.0
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0
torch>=2.0.0
